Fix missing biodiversity layers after relad
